Remove compile warning for OverlayScalarMapperLegend
* apply explicit unsigned int -> float cast.

+++ Fwk/AppFwk/cafVizExtensions/cafOverlayScalarMapperLegend.cpp
@@ -185,7 +185,7 @@ void OverlayScalarMapperLegend::renderGeneric(OpenGLContext* oglContext, const V
 
     this->computeLayoutAndExtents(position, size);
 
-    Vec2f backgroundSize(this->matchedWidth(), (float)size.y());
+    Vec2f backgroundSize((float)this->matchedWidth(), (float)size.y());
 
     // Do the actual rendering
     if (software)
